+++ tests/alchemyapp/models.py
@@ -5,8 +5,7 @@
 import os
 
 from sqlalchemy import Column, Integer, Unicode, create_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import scoped_session, sessionmaker
+from sqlalchemy.orm import declarative_base, scoped_session, sessionmaker
 
 try:
     import psycopg2  # noqa: F401